pyproject.toml
[tool.poetry]
name = "source-hub"
version = "0.1.0"
description = ""
authors = ["Matthew Brown <mbrown@bhei.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.75.0"}
python-decouple = "^3.6"
passlib = "^1.7.4"
motor = "^2.5.1"
python-jose = "^3.3.0"
beanie = "^1.10.1"
bcrypt = "^3.2.0"
jsonpatch = "^1.32"
typer = "^0.4.1"
playwright = "^1.20.1"
aiohttp = "^3.8.1"
aiofiles = "^0.8.0"
pdfplumber = "^0.6.0"
python-dateutil = "^2.8.2"
boto3 = "^1.21.42"
tenacity = "^8.0.1"
xxhash = "^3.0.0"
playwright-stealth = "^1.0.5"
redis = "^4.2.2"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
boto3-stubs = {extras = ["essential"], version = "^1.21.42"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
